refactor(ui): move renderTeamSync to team-sync/render/render-team-sync.ts, barrel the tab
Finish the team-sync.ts decomposition. renderTeamSync (plus the
TEAM_SYNC_ACTIONS_MOUNT_ID const and the teardownTeamSyncRender helper
only used here) moves into render/render-team-sync.ts. team-sync.ts
becomes a thin barrel re-exporting the four public entrypoints
(setLoadSyncData, initTeamSyncEvents, renderTeamSync,
renderSyncSharingReview) so index.ts wiring stays on the same import
paths.
